Change how `shell` works and make activation more robust (#316)
For `shell`, we used to run the activation script and then start the
shell.
The problem was that the shell also runs a `bashrc`, `zshrc` or similar
file after starting. So any PATH modifications (or env var
modifications) we had done might have been un-done by the `bashrc` file.

This PR changes how `shell` works. It first starts the shell (that runs
all `bashrc` etc. commands) and then sources an activation script.
This is only possible with some UNIX foo (forking and mirroring the
shell).

We copied the approach from `poetry` that uses `pexpect`. We forked
parts of `rexpect` here to make it work.

---------

Co-authored-by: Bas Zalmstra <zalmstra.bas@gmail.com>
Co-authored-by: Ruben Arts <ruben.arts@hotmail.com>