feat: install pypi dependencies (alpha) (#452)
This is an initial version to support pypi dependencies (#175).

There is still lots of functionality missing but the main implementation
works. Just add:

```toml
[pypi-dependencies]
plotly = ">=5.18.0" # or whatever
```

to your `pixi.toml` and enjoy.

This work builds on https://github.com/prefix-dev/rip which means
`sdists` are currently not (yet) supported.