feat: implement lock-file satisfiability with pypi dependencies (#494)
This PR adds support for checking the satisfiability of the lock-file
which includes pypi-dependencies.

Purls have been added to the lock-file
(https://github.com/mamba-org/rattler/pull/414) (See also:
https://github.com/conda-incubator/ceps/pull/63). This enables checking
which conda packages will install which pypi packages without needing to
check the internet. This ensures we can still check if a lock-file is up
to date quickly.

I did not profile this code but I think there are a lot of places we can
improve the performance. Thats for a later PR.

I also didn't add tests. I think we should but we can also do that in
another PR.

Closes #467

---------

Co-authored-by: Ruben Arts <ruben.arts@hotmail.com>