feat: use range when adding requirement (#536)
Implements the behavior described in #285 .

When running `pixi add python` we would previously look at all
platforms, add the requirement, and solve the environment. We would then
take the lowest version between all platforms and pin that. So currently
for python that would result in the requirement:

```toml
python = "3.12.0.*"
```

This is a very strict requirement. And would also fail to resolve if
this version (or a higher version) is not available for all platforms
(see #526 ).

This PR changes the behavior by instead adding a range requirement that
allows change in the last segment of the version. And allowing for
multiple versions if a version is not available for all platforms.

For python this would become:

```toml
python = ">=3.12.0,<3.13"
```

For the issue described in #526, this becomes:

```toml
ros-humble-rosbridge-server = ">=1.2.0,<1.4"
```

Note that `1.4` is selected instead of `1.3`. This is because `1.2` is
[not available for all
platforms](https://prefix.dev/channels/robostack-humble/packages/ros-humble-rosbridge-server).

Fix #526 
Fix #285