Multi env proposal documentation (#584)
This proposal came out of a discussion between developers from
prefix.dev, QuantCo and Anaconda.

In this discussion we went over multiple use cases and came to the
conclusion that this approach would give us a good cover of all use
cases we could think of.

This is a design proposal and not a set in stone requirement. We will
iterate on the design while implementing it.

Feel free to give feedback!

This will be visible in the documentation website, this is also the best
way to read it as I used Mkdocs styling to improve readability. (`pixi
run docs` of course :wink: )

Ping: @pavelzw @0xbe7a @travishathaway

---------

Co-authored-by: Pavel Zwerschke <pavelzw@gmail.com>
Co-authored-by: Bela Stoyan <bela.stoyan@gmail.com>
Co-authored-by: Travis Hathaway <travis.j.hathaway@gmail.com>