feat: add windows, macos, linux and unix to targets (#832)
I added a few more targets that should make it easier to select packages
for some sets of platforms.

You can now do:

```toml
[target.unix.dependencies]
special-unix-pkg = "*"

[target.windows.dependencies]
pywin32 = "3.12"

# ... linux and macos are also available.
```

We could consider to overload `osx`, too. The correct name for osx is
macOS though, since a while. Unfortunately that is not reflected in the
"subdir" name.