feat: add error for unsupported pypi dependencies (#1052)
Fixes #1036

With the following pixi toml:

```toml
[project]
name = "project"
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64"]

[dependencies]
python = "3.11.8"

[pypi-dependencies]
timeago = "==1.0.16"

[feature.test]
platforms = ["linux-64"]

[environments]
test = ["test"]
```

The environment `test` is only solvable for `linux-64`. However, when
trying to solve this lock-file on another platform no python interpreter
can be installed for the environment because the environment does not
support it.

This PR shows a problem that we should fix, see #1051 for that tracking
issue.