feat: tree command
fixes: #238

Adds a tree command for the CLI that can print out the tree of dependencies based on those defined in pixi.toml. Additionally it can filter the regular dependency tree, or show what tree of packages need a given package.

It keeps track of visits for the regular tree, but currently does not for the inverted one.