fix: unsolvable when moving dependency from conda to pypi (#1099)
Fixes #1098 

When a package was moved from conda dependencies to pypi dependencies
and the version is changed the satisfiability reported that a
requirement was satisfiable. However, if a pep508 requirement isn't
satisfiable, only the pypi packages are updated, but in this case the
conda packages should also be updated.

To fix this I added a separate error for a requirement mismatch on a
conda package. This will trigger updating the conda packages as well.