fix: allowing solving for different platforms (#1101)
Fixes #1082

Checking if the current machine meets the project requirements is only
really required when using `run` or `shell`. This PR moves the logic to
check if the system meets the minimum system requirements to those
specific CLI commands. This allows you to create lock-files for other
platforms than your own. (as long as you dont have pypi dependencies)

Co-authored-by: Ruben Arts <ruben.arts@hotmail.com>