fix: editable non-satisfiable (#1251)
Fixes an issue where multiple requirements of editable/non-editable
would conflict and cause the lock-file to be non satisfiable. However,
the logic should be that if just one requirement specifies an editable
install then it should be editable regardless of what other requirements
specify.

You also get a nice error message:

```
expected bar, and foo to be editable but in the lock-file they are not, whereas baz is NOT expected to be editable which in the lock-file it is
```

Note that I also included several tests that test several cases.