feat: alternative pypi registries (#1225)
Enables the use of alternative PyPI registries. The definition of these
can be specified per feature. There are certain rules for making unions
out of these that are explained throughout the code.

This adds the options for:
* extra-index-urls
* index-urls
* find-links

Which follows the familiar pypi conventions. These are deserialized and
passed per environment or solve-group solve to uv.

---------

Co-authored-by: Bas Zalmstra <zalmstra.bas@gmail.com>
Co-authored-by: Bas Zalmstra <bas@prefix.dev>
Co-authored-by: Ruben Arts <ruben.arts@hotmail.com>