feat: change global config to kebab-case
This PR adds the option to use kebab-case in the global config.
Which will be the preferred option, for older options we've kept the
old way of `snake_case` configuration.

Also, the documentation was updated to reflect this. There was as error
in the `repodata-config` in the documentation this was incorrectly named
`repodata-options` this has *never* worked. This has also been
rectified.