feat: add env to activation table (#1156)
> [!NOTE]
> I'm not a 100% happy with the fact that this doesn't work with
environment variables that already exist. For example:
> I would expect it to expand variables if you would do:
> ```toml
> [activation.env]
> PATH = "bla/bli/blu:$PATH"
> PATH_TO_XACRO =
"$PIXI_PROJECT_ROOT/.pixi/envs/default/share/ros/bla/bli/xacro.xacro
> ```
> But like the scripts, this also isn't cross platform/shell. 

That said we made the decision to clean this up and release it as a
first version so users that can use it will benefit from it and it would
give us more time to come up with a good solution.