fix: too much shell variables in activation of `pixi shell` (#1507)
Fixes #1503

The issue was that the work on the clean-env introduced a situation
where all env vars of the current environment are added to the shell
activation. This is not what we want, so this pr reverts that behavior.

This pr adds
- Cleanup of the activation code
- Storage of both the clean and normal environment variables when used
in the tasks.
- Introduces Environment variable behavior to tell the activator what to
do with the current environment
- Removes windows non-sense, I added the possible code to the related
issue : #289 for future reference.