feat: bump system requirements (#1553)
This PR bumps the system requirements to be more in line with current
`manylinux2_28` and minimum macOS versions.

While conda-forge is very conservative with the default system
requirements, we can be a bit more bold. We based our first set of
defaults on `conda-forge` defaults, but the aims are somewhat different:
conda-forge tries to build packages for maximum compatibility while we
are developing a software that is modern and probably mostly used on
recent machines.

There is some pain with wheels that are only available for macOS 13.0+
which have become somewhat common on PyPI. Users would then receive
`sdists` with all of their problems. Therefore I think it would be good
to bump the minimum here. Users that want to support older systems or
RedHat style Linux versions can always set different values in the
`pixi.toml`.