feat(manifest): change pypi-options to function more like channels and platforms (#1405)
In the #1385 we've decided to give this a try.

This PR makes `pypi-options` work more like channels and platforms. They
can now not be specified for the default feature so it's more like:

```toml
[project]
name = "pypi-custom-registry"
authors = ["Tim de Jager <tim@prefix.dev>"]
channels = ["conda-forge"]
platforms = ["osx-arm64", "osx-64", "linux-64", "win-64"]

[project.pypi-options]
index-url = "https://pypi.org/simple"
```
In addition, they can still be used within the default feature if need be.
In a subsequent PR, we can add the ability to use [feature.default.channels] etc. (now you cant do this) so that it functions similarly to pypi-options in the default feaure.


---------

Co-authored-by: Ruben Arts <ruben.arts@hotmail.com>