fix: flaky task test (#1581)
Fix for flaky task test, not using `remove_var` because of:
https://doc.rust-lang.org/std/env/fn.remove_var.html


> Even though this function is currently not marked as unsafe, it needs
to be because invoking it can cause undefined behaviour. The function
will be marked unsafe in a future version of Rust. This is tracked in
[rust#27970](https://github.com/rust-lang/rust/issues/27970).

> This function is safe to call in a single-threaded program.

> In multi-threaded programs, you must ensure that are no other threads
concurrently writing or reading(!) from the environment through
functions other than the ones in this module. You are responsible for
figuring out how to achieve this, but we strongly suggest not using
set_var or remove_var in multi-threaded programs at all.