docs: Add changelog to docs (#1574)
### Summary

Add a hook to copy the changelog to the docs build so that it's
available via the site too. I was frequently swapping between github and
the website to see the changes in the new versions.

Also updated is the exitsting changelog itself so that the TOC is
rendered properly in the built site. See preview below


![image](https://github.com/prefix-dev/pixi/assets/62500639/d9d2bf67-e65a-49c3-9572-9a4294d6e327)

In addition, the template used to generate the documentation has also
been modified slightly so that the future docs are rendered with the
updated heading sections.

---------

Co-authored-by: Ruben Arts <ruben@prefix.dev>