feat: Add nushell completion
- Add nushell completion
- Stop guessing shell, and make `--shell` required
- Adapt instructions

I have a version local that still tries to guess the shell from the `$SHELL` env var.
But it made the code much more verbose, and I am not a fan of that behavior.
On my system `$SHELL` gives `/bin/bash` even though I use nushell.

I also adapted the instructions to stop suggesting that your choice of shell is platform dependent.
Every listed shell can be used on Windows and vice-versa.