feat: use locked pypi indexes during installation (#1670)
When thinking about how install should work off the lock file mostly, I
found out that we are using the manifests pypi-options in the
installation. I think this can be replaced by the locked indexes,
especially as we have the actual url's at hand when making the `Dist`'s.

This implements that refactor, the only assumption I've made is that the
first url in the list is the equivalent of the `IndexLocation`
`index_url`, this mainly being the main pypi index in most cases.