docs: fix mlx feature in "multiple machines" example
The example as it was would not solve, it ends with a fairly
obscure error:

```
  × failed to solve the conda requirements of 'mlx' 'osx-arm64'
  ╰─▶ Cannot solve the request because of: mlx >=0.5.0,<0.6.0 cannot be installed because there are no viable options:
      └─ mlx 0.5.0 | 0.5.0 | 0.5.0 | 0.5.0 | 0.5.0 | 0.5.1 | 0.5.1 | 0.5.1 | 0.5.1 | 0.5.1 would require
         └─ __osx >=13.3, for which no candidates were found.
```

This was already found previously in
https://github.com/prefix-dev/pixi/issues/562#issuecomment-1855847020,
and the fix is to add a minimum macOS version.
Note that the conda-forge package metadata (saying >=13.3) is also incorrect,
the 13.5 minimum added to the example comes from the MLX docs:
https://ml-explore.github.io/mlx/build/html/install.html#

While we're at it, also update to a recent MLX version.