fix: use two in memory indexes, for resolve and builds (#1969)
This PR makes the resolver use two memory indexes, one for the
resolution process and one for the potential building of wheels. Even
though uv uses different internal revolvers it was sometimes still
getting data that was overridden by us during wheel builds, which was
unexpected.

This was cropping up in issues like #1686. We've made a minimal version
of the lock file where this problem was occuring. And added that as a
test.

---------

Co-authored-by: Hofer-Julian <30049909+Hofer-Julian@users.noreply.github.com>