feat: move matrix to workflow that it is used in (#1987)
Move the matrix to the workflows that it is used in, there was no
current advantage of it being split. One idea was that it could be
triggered per finished build per os, but as it was depending on build
this was not happening. And it is not possible to trigger an entire
workflow from a step.