feat: Export conda environment.yml
Builds upon #1873 to add exporting of `environment.yml` files.

Since #1873 provides the conda-lock style matrix of exports, I focused on exporting an environment.yml that matches the manifest.

Tests against a bunch of the different example manifests for the various pypi-options, and I tested locally that at least some of those were installable with micromamba.

Replaces #1427