Make self-update a compile time feature
Currently pixi refuses to self-update if it is not installed at its
default location. This can be overridden using the `--force` flag.

This is not ideal for two reasons: Some people might intentionally
install pixi to a non-default location, such as `~/.local/bin/pixi`,
while not using a package manager. On the other hand, passing `--force`
still allows pixi installations installed by a package manager to be
broken.

A better solution would be disabling the self-update functionallity at
compile time. This PR adds a compile-time feature called
`self_update`, which is not enabled by default.
This PR also enables this feature in the CI build so releases produced
by CI can still be self-updated. Because the feature is not enabled by
default, other distributions of pixi, such as when built by other package
managers, will not support self-update.

Instead of completely removing the self-update sub-command, disabling
this feature only replaces it with a stub command that outputs an error
message. This is to prevent user-confusion, when users read about
self-update in documentation, only to find the sub-command missing in
their installation.

With this implemented the default location check and `--force` flag will
not be needed anymore.