chore: bump Rust toolchain to 1.81.0 (#2227)
The Rust 1.81 toolchain fixed the hashing of Path objects. Previously,
paths were hashed equally wether they contained a slash or not. This is
not the case anymore which is definitely fixing a bug.

However, it does break existing hashes in the task cache or in lockfiles
for editable installs, which is unfortunate for us.

---------

Co-authored-by: Tim de Jager <tim@prefix.dev>