Create `.pixi/.gitignore` containing `*`
This has the advantage that `.pixi/` will be ignored in a self-contained way
without requiring modification of the project-level `.gitignore`.

This is mostly redundant with the rule defined in `GITIGNORE_TEMPLATE`, but
retaining them both for the moment ensures that edge cases are better handled.
See the discussion in <https://github.com/prefix-dev/pixi/issues/2115>.