chore: update to latest rattler/feat/pixi-build (#2500)
This PR updates pixi to be compatible with the latest changes in the
`feat/pixi-build` branch of `rattler`. Once this PR is ready to be
merged we can merge `feat/pixi-build` into `main` in the `rattler`
repository. We can then quickly after that merge `feature/pixi-build` of
this repo into `main` too.