docs: adapt completion instructions
We currently give people a one-liner to setup completions.

That is convenient, but also fragile.
If people's startup script doesn't end in a newline, it [breaks](https://github.com/panel-extensions/copier-template-panel-extension/issues/4#issuecomment-2495774437)
If people run it multiple times it will be duplicated.

Let's follow starship.rs and give users instructions to add it themselves