feat: pixi rattler build (#2526)
## Overview:
In this PR we add rattler-build as a new backend. This means that user
can build a `recipe.yaml` if it is next to a`pixi.toml` or in `recipe`
folder that is nearby a `pixi.toml`.
It is somehow similar with current existing `conda-build` backend, but
instead of calling it directly and having implementation done on pixi
side, we lift all work on pixi-build-backend side using jsonrpc (
similar with pixi-build-python ).