feat: add package specific dependencies and errors (#2556)
Builds on https://github.com/prefix-dev/pixi/pull/2544 .

This PR adds dependencies to either workspace or package targets based
on the rest of the content of the manifest.

E.g.

```toml
[project]

[host-dependencies]

[dependencies]
```

Adds both the `host-dependencies` and the `dependencies` to the default
target of the default feature of the workspace/project. But:

```toml
[workspace]
preview = ["pixi-build"]

[dependencies]

[package]

[host-dependencies]

```

Adds the `host-dependencies` to the default target of the package
manifest instead.

I also added _loads_ of corner error cases with helpful snippets and
help messages to clarify what happens when enabling the preview feature
(see the pushed snippets).

Build backends should read the package section of the manifest and
disregard the workspace/project section. But thats for follow-up PRs to
the backends.