refactor: `[build-system.build-backend]` (#2601)
### Description 

We decided to refactor the way we format the `[build-system]` section
because it was confusing, inconsistently formatted with the rest of the
file and contained some duplication.

Previously the `[build-system]` section would look like:

```toml
[build-system]
dependencies = ["python-build-backend > 12"]
build-backend = "python-build-backend"
channels = ["conda-forge"]
```

We decided to reformat this as:

```toml
[build-system]
build-backend = { name = "python-build-backend", version = ">12"  }
channels = ["conda-forge"]

# optional
additional-dependencies = { git = "*" }
```

The `build-backend` describes both the package and the executable to
invoke in the package. Additional dependencies can be added through the
`additional-dependencies` field. I also added loads of errors for corner
cases.