fix: color formatting detection on stdout (#2613)
Text printed to stdout by pixi has colors even when stdout is redirected
to a pipe or file. For example:

> $ pixi list | xxd -l 64
> 00000000: 1b5b 316d 5061 636b 6167 651b 5b30 6d20  .[1mPackage.[0m

It is a common feature for command-line tools to automatically detect if
stdout is a pipe or file and disable colors. And in fact pixi uses the
crate `console` for color formatting which has built-in support for
detecting if stdout and stderr are terminals and automatically disabling
color output, so this should work.

However pixi does its own color management which overrides how `console`
manages color styling. Pixi checks if _only_ stderr is a terminal, and
based on that overrides `console` handling of stdout and stderr. Thus
causing stdout to be color formatted when stdout is redirected to a file
but stderr isn't.

Since `console` already has detection for both stdout and stderr, lets
default to that. We only override `console` when `pixi --color`,
PIXI_COLOR, or NO_COLOR are set.

As a bonus, this means pixi will now support the `CLICOLOR` spec that is
implemented by `console`.

Fixes: #2536