feat: make binary, config folder, and lock file names dynamic
Make the pixi binary name dynamic so that it's determined off the actual binary
name when invoked so if it's renamed, it reports correctly.

Make the pixi config folder and default lock file to use overridable in cargo config.

Make the default channels overriable in a build in cargo config. Automatically parse
full URLs in default channels list as URLs and not short names at compile time.

Make some functions and mods accessed externally pub instead of pub(crate).