fix: use correct directory for build artifact cache (#2830)
There was an issue were the packages in the build-cache were reused for unrelated source checkouts. This is because we bucket the caches based on the pinned source spec but for a lot of examples the pinned path source spec was simply .. This caused a lot of examples to reuse the cache entry of another source checkout causing strange behavior.

I fixed this issue by using the pinned source spec as the cache key if the source is immutable (e.g. for git checkouts) and otherwise (e.g. for path source dependencies) use the absolute path on disk as the cache key. I think this makes sense as it creates separate cache entries for packages in different locations on disk (like temporary directories).