feat: move pixi build type conversions into its own crate (#2866)
This moves the conversion methods used in the pixi_build_frontend to currently extract the PackageManifest into the appropriate build types into its seperate crate.

Why? Because we need an implementation of this in the backend CLI (which emulates being a frontend) but do not require the rest of the pixi frontend code.