test: add `online_tests` feature to control Internet use
Add an `online_tests` feature that specifically covers tests accessing
the Internet.  It is enabled by default to preserve the existing
behavior, but can be explicitly disabled when running the test suite
entirely offline.

I have not added the respective feature checks to
`slow_integration_tests`, as they all seem to require Internet, so that
seems redundant.

Fixes #2880