refactor: spawn main entrypoint in box (#2892)
When running a debug build on Windows, you often get "stack overflow" errors when starting pixi. This is because the default stack size on Windows is smaller than on other operating systems. This PR works around this issue by boxing the main entry point future instead, which effectively moves it to the heap.