fix: stackoverflow when running pixi in debug mode on windows (#2922)
Stack overflow issues are mostly caused by the fact that windows has a 1MB stack size for the main thread. Which is relatively small to the 8MB on macOS and Linux.

Apparently, there is a not very well-documented environment variable to set the stack size of threads called RUST_MIN_STACK. But this only sets the stack-size for newly spawned threads, while the issues happens on the main thread. I noticed uv very recently worked around this issue to spawn an additional main thread from which the entire application is ran. Pretty clever, so I implemented the same behavior for pixi.