fix: shell-hook `--no-completion`
Fixes #3549

Merging of config was faulty for `pixi shell-hook`, this PR fixes that.
Unfortunately, this also includes a breaking change. Seems like we missed that `Option<bool>` leads to a bit weird CLI which requires users to type `--no-completion true`, which doesn't make a lot of sense. Changing to `bool`, changes this to `--no-completion`
Since this is breaking anyway, I also adapt `--no-completion` to `--no-completions`.