fix: hashing for same package-name (#3775)
I think it is more correct: to hash and compare on the normalized package name. Now this issue is not causing an error while it should: #3736

This was because passing in the normalized name to a function that checks if that PyPI dependency existed for that target would not return true if the source names do not match up.