feat: support legacy mojoproject.toml
As we sunset Magic at Modular and push folks to use Pixi instead, this feature silently adds support for the legacy mojoproject.toml format to keep existing legacy projects working as we transition off.

mojoproject.toml was always a superset of pixi.toml that worked with older tools at modular.