fix: panic on moving source location
The problem was that `reporter_context` when solving pixi environments, assumes that `self.solve_pixi_environments[id]` is present.
This doesn't seem to be guaranteed if one of the solves fails, in our test case we had an invalid path for a source dependency.

By checking if there's actually an entry for `id` we avoid panicking during indexing