fix: recursive optional dependencies (#4743)
This changes the mechanism of feature building with pyproject.toml, and appends the features together instead of overwrites. This enables recursive optional dependencies and dependency-groups, that is feature inheritance. So before when encountering a duplicate requirement we would error, but now we append.
 
Also two unit tests were added to make sure this works on a manifest level. A manual test was done on the original issue: #4706